Split up BGAPI tests into separate modules for clarity.
